Separate part of GrabState as another class for testability
